added the words "by" and "piece" to the paragraph, corrected spelling of "occupied"
Your `Piece` will need to (1) track its position and (2) hold a
reference to the `Board`. The `SlidingPiece` in particular needs the
`Board` so it knows to stop sliding when blocked by another
piece. Don't allow a piece to move into a square already occupied by the
same color piece, or to move a sliding piece past a piece that blocks it.